test: Comprehensive test suite updates and improvements
Common test infrastructure (tests/common/):
- artifacts.rs: Updated artifact generation helpers
- input_sim.rs: Input simulation improvements
- mock_terminal.rs: Better terminal mocking patterns

E2E tests:
- e2e_input_flow.rs: Streamlined input flow testing
- e2e_perf.rs: Significant refactoring for cleaner benchmarks
- e2e_render_cycle.rs: Improved render cycle assertions
- e2e_unicode.rs: Enhanced Unicode handling tests

Property tests:
- proptest_buffer_diff.rs: Buffer diff property tests
- proptest_color_ops.rs: Color operation properties
- proptest_text_unicode.rs: Unicode text properties

Other tests:
- fixtures_test.rs: Fixture validation
- golden_rendering.rs: Golden file comparisons
- pty_e2e.rs: PTY integration tests
- repro_issues.rs: Issue reproduction tests

+++ tests/common/input_sim.rs
@@ -502,7 +502,7 @@ pub fn key_to_ansi(key: &KeyEvent) -> Vec<u8> {
                 10 => 21,
                 11 => 23,
                 12 => 24,
-                _ => 15 + n as u8,
+                _ => 15 + n,
             };
             result.extend_from_slice(format!("\x1b[{code}{mod_suffix}~").as_bytes());
         }
